flatten: smart_text widget hierarchy
