+++ packages/client/lib/app/core/widgets/smart_text/constants/messages_data.dart
