+++ packages/client/lib/app/core/widgets/smart_text/movies/movies.dart
