+++ packages/client/lib/app/core/widgets/smart_text/movies/static_smart_text_color_movie.dart
