--- packages/client/lib/app/core/widgets/smart_text/stack/constants/constants.dart
@@ -1,3 +1,0 @@
-export 'types/types.dart';
-export 'movies/movies.dart';
-export "data/data.dart";
