--- packages/client/lib/app/core/widgets/smart_text/stack/constants/types/types.dart
@@ -1,3 +1,0 @@
-export "./fading_status.dart";
-export "./gestures.dart";
-export "./rotating_text_data.dart";
