+++ packages/client/lib/app/core/widgets/smart_text/types/fading_status.dart
