+++ packages/client/lib/app/core/widgets/smart_text/types/gestures.dart
