+++ packages/client/lib/app/core/widgets/smart_text/types/rotating_text_data.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:nokhte/app/core/widgets/widget_constants.dart';
+import 'package:nokhte/app/core/widgets/widgets.dart';
 
 // ignore: must_be_immutable
 class RotatingTextData extends Equatable {
