+++ packages/client/lib/app/core/widgets/smart_text/types/types.dart
@@ -1,0 +1,3 @@
+export "fading_status.dart";
+export "gestures.dart";
+export "rotating_text_data.dart";
